improve gfile.CopyDir function; add Model function for gdb.Model
