+++ .example/other/config.toml
@@ -1,4 +1,7 @@
 
+[database]
+	debug = true
+    link  = "mysql:root:12345678@tcp(127.0.0.1:3306)/test"
 
 [redis]
     default  = "127.0.0.1:6379,0"
