+++ .example/other/test.go
@@ -1,17 +1,13 @@
 package main
 
 import (
-	"github.com/gogf/gf/net/ghttp"
+	"github.com/gogf/gf/frame/g"
 )
 
 func main() {
-	s := ghttp.GetServer()
-	s.BindHandler("/log/path", func(r *ghttp.Request) {
-		r.Response.Writeln("请到/tmp/gf.log目录查看日志")
-	})
-	s.SetLogPath("/tmp/gf.log")
-	s.SetAccessLogEnabled(true)
-	s.SetErrorLogEnabled(true)
-	s.SetPort(8199)
-	s.Run()
+	db := g.DB()
+	// 开启调试模式，以便于记录所有执行的SQL
+	db.SetDebug(true)
+
+	db.Table("user").Delete("score < ", 60)
 }
